First steps to getting callbacks with new data
Wrap each device type's data in a struct to easily pass it back to the
callback.

Define a mechanism to trigger callbacks on the base device and add
device specific implementation for callback delegate etc.

Update tests to respect the new location of the polled data. Include
NUnit NuGet package as part of the update.

TODO: Define a more general scheme for callbacks somehow. There is a lot
of repeated boilerplate to set up the callbacks per device, and there is
no API for registering a callback specified in abstract on the base
device.