Use fixed buffers for response/channel event callbacks
This commit enforces `fixed` buffers for any buffer given to the ANT+ C
library, for it to write to on occurrence of an event.
Garbage collection in the C# VM may relocate variables during standard
operation. C# provides the `fixed` keyword to prevent this from
happening for static buffers.

There was a bug, where the heart rate would cease to update after a
certain time. It was found that, after garbage collection ran, the
buffer was no longer being written to before the CE callback function
was called. The root cause of this was that the byte array being used
as a buffer was being relocated during the GC run, and the C library
making the callback was (obviously) unaware of this change.